requirements.txt
appnope==0.1.2
azure-core==1.10.0
azure-identity==1.5.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
decorator==4.4.2
idna==2.10
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
msal==1.8.0
msal-extensions==0.3.0
msgraphcore==0.0.2
pexpect==4.8.0
pickleshare==0.7.5
portalocker==1.7.1
prompt-toolkit==3.0.11
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
requests==2.25.1
six==1.15.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
